requirements.txt
tinydb==4.7.1
python-dotenv==1.0.0